requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
bottle==0.12.19
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Flask==2.0.3
fonttools==4.29.1
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pygame==2.1.2
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
telepot==12.7
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2